* Host-Compute: implemented full subgroup support:    * host_function/elf_binary: we now keep track of the subgroup index and subgroup lane index as well (instance_ids_t) + added restore() helper function for external dynamic ID restore    * host_function/host_group: implemented all subgroup shuffle operations for 16-/32-bit integer+float scalar and vector types    * host_function: extended simd_storage to have space for all possible data types (16 bytes at most)    * host_function: implemented SIMD barrier functionality -> with this, we will only switch between fiber contexts within the same SIMD group (i.e. 0 -> 1 -> 2 ... -> 31 -> 0)    * host_function: unified the x86 and ARM fiber context: both now use a stack size of 16KiB (with the stack aligned to 16KiB as well!) and both contexts are now 256 bytes in size    * host_function: fully drop ucontext support, we no longer need or want this    * host_function: each floor_fiber_context now stores its associated local-linear index, sub-group index, sub-group-local index and global/local index -> this simplifies context switching and ID restore + allows for more introspection    * host_function: each floor_fiber_context now also stores the index to the next linear/work-group context and SIMD context in line (again, faster switching)    * host_function: added floor_current_context() helper function that returns the current floor_fiber_context based on the current stack pointer (as we already store a pointer to the floor_fiber_context in each fibers stack)    * host_function: device_exec_context_t and host_exec_context_t now inherit from a common base class    * device Host-Compute group: added all subgroup shuffle, reduce and scan operations    * device Host-Compute: simd_barrier() now has its own implementation + only performs the barrier with the SIMD group    * NOTE: this generally seems to work on x86-64, but on AArch64 I still have occasional issues in complex compute kernels  * aligned_ptr: it's now possible to request an alignment that is a multiple of the page size in make_aligned_ptr()  * device (+all backends): removed indirect_command_support and indirect_compute_command_support feature flags+testing as this is universally supported now (toolchain + device_info retain the feature flags for now)  * toolchain: Host-Compute AArch64 code is now always built with -mno-unaligned-access  * toolchain: added Metal "restrictive_vectorization" build option (off by default): if enabled, this restricts the amount of scalar->vector transformations that can be made and only creates up to 4D vector types (this puts it on the same level as Vulkan)  * device algorithm/group: moved the min_op/max_op definitions to group.hpp + use these in the individual backends as well now  * device algorithm: added a "prefer_simd_operations" flag/constant that determines whether SIMD optimized algorithms should be used for a backend (enabled by default for all backends except Host-Compute)  * device common: moved the general group.hpp to before the backend specific group includes  * device common: local_buffer now uses alignas() rather than the aligned attribute + is now explicitly marked as used  * CUDA group: use actual lane ID register rather than computing it ourselves in cuda_sub_group_scan()  * CUDA ID: fixed signedness of sub_group_local_id and sub_group_count  * Metal: fixed simd_barrier(), this now actually calls air_simdgroup_barrier() ...  * host_context/host_function: host_context now initializes host_function on construction, determining and storing the logical CPU core count once on startup rather than every time at run-time (faster + fixes the issue that only one core is determined to be available when using thread affinity)  * metal_program: improved/fixed determining if a compute kernel can be indirect (now checks for non-global-AS parameters, thus failing on param<>/constant as well now)  * metal_program: in debug mode, print whether a compute kernel has indirect support as well now  * version bump to v0.5.0-a3

+++ include/floor/device/backend/cuda_id.hpp
@@ -60,9 +60,9 @@
 
 FLOOR_SUB_GROUP_ID_RANGE_ATTR const_func uint32_t get_sub_group_id() asm("floor.get_sub_group_id.i32");
 #define sub_group_id get_sub_group_id()
-#define sub_group_local_id __nvvm_read_ptx_sreg_laneid()
+#define sub_group_local_id uint32_t(__nvvm_read_ptx_sreg_laneid())
 #define sub_group_size FLOOR_DEVICE_INFO_SIMD_WIDTH
-#define sub_group_count __nvvm_read_ptx_sreg_nwarpid()
+#define sub_group_count uint32_t(__nvvm_read_ptx_sreg_nwarpid())
 
 FLOOR_GLOBAL_ID_RANGE_ATTR const_func static uint32_t get_global_id(uint32_t dim floor_unused) FLOOR_CUDA_DIM0 {
 	return __nvvm_read_ptx_sreg_ctaid_x() * __nvvm_read_ptx_sreg_ntid_x() + __nvvm_read_ptx_sreg_tid_x();
